Use np.float64 for classes_

+++ python/treelite/sklearn/exporter.py
@@ -162,14 +162,17 @@ def _export_tree(
             subestimator_state.update(
                 {
                     "n_classes_": n_classes[0],
-                    "classes_": np.arange(n_classes[0]),
+                    "classes_": np.arange(n_classes[0], dtype=np.float64),
                 }
             )
         else:
             subestimator_state.update(
                 {
                     "n_classes_": n_classes.tolist(),
-                    "classes_": [np.arange(n_classes[i]) for i in range(n_targets)],
+                    "classes_": [
+                        np.arange(n_classes[i], dtype=np.float64)
+                        for i in range(n_targets)
+                    ],
                 }
             )
     subestimator.__setstate__(subestimator_state)
